Fixing Docs.rs build fail (#8)

+++ build.rs
@@ -1,7 +1,7 @@
 use flate2::read::GzDecoder;
 use std::error::Error;
-use std::fs::{remove_file, File};
-use std::io::Cursor;
+use std::fs::{self, remove_file, File};
+use std::io;
 use std::path::Path;
 use tar::Archive;
 
@@ -11,39 +11,53 @@ const EIGEN_TAR_NAME: &str = "eigen-3.4.0.tar.gz";
 
 type GenericError = Box<dyn Error>;
 
-async fn get_file_from_url<P>(url: &str, path: &P) -> Result<(), GenericError>
+fn get_file_from_url<P>(url: &str, path: &P) -> Result<(), GenericError>
 where
     P: AsRef<Path>,
 {
-    let response = reqwest::get(url).await?;
-    let mut file = std::fs::File::create(path)?;
-    let mut content = Cursor::new(response.bytes().await?);
-    std::io::copy(&mut content, &mut file)?;
-    Ok(())
+    use curl::easy::Easy;
+    use std::io::Write;
+
+    let f = fs::File::create(path)?;
+    let mut writer = io::BufWriter::new(f);
+    let mut easy = Easy::new();
+    easy.url(url)?;
+    easy.write_function(move |data| Ok(writer.write(data).unwrap()))?;
+    easy.perform()?;
+    let response_code = easy.response_code()?;
+
+    if response_code == 200 {
+        Ok(())
+    } else {
+        Err(format!("Unexpected response code {} for {}", response_code, url).into())
+    }
+    // Ok(())
 }
 
-#[tokio::main]
-async fn main() -> Result<(), GenericError> {
-    let out_dir = std::env::var("OUT_DIR")?;
-    let out_path = Path::new(&out_dir);
-    let eingen_tar_path = out_path.join(EIGEN_TAR_NAME);
+fn main() -> Result<(), GenericError> {
+    if !std::env::var("DOCS_RS").is_ok() {
+        // ... your code here ...
 
-    let eigen_dir = out_path.join(EIGEN_PATH);
-    if !eigen_dir.is_dir() {
-        get_file_from_url(URL, &eingen_tar_path).await?;
-        let tar_gz = File::open(&eingen_tar_path)?;
-        let tar = GzDecoder::new(tar_gz);
-        let mut archive = Archive::new(tar);
-        archive.unpack(out_path)?;
-        remove_file(eingen_tar_path)?;
+        let out_dir = std::env::var("OUT_DIR")?;
+        let out_path = Path::new(&out_dir);
+        let eingen_tar_path = out_path.join(EIGEN_TAR_NAME);
+
+        let eigen_dir = out_path.join(EIGEN_PATH);
+        if !eigen_dir.is_dir() {
+            get_file_from_url(URL, &eingen_tar_path)?;
+            let tar_gz = File::open(&eingen_tar_path)?;
+            let tar = GzDecoder::new(tar_gz);
+            let mut archive = Archive::new(tar);
+            archive.unpack(out_path)?;
+            remove_file(eingen_tar_path)?;
+        }
+
+        cc::Build::new()
+            .cpp(true)
+            .include(eigen_dir)
+            .file("src/solver.cpp")
+            .flag_if_supported("-std=c++1y")
+            .compile("solver_cpp");
     }
-
-    cc::Build::new()
-        .cpp(true)
-        .include(eigen_dir)
-        .file("src/solver.cpp")
-        .flag_if_supported("-std=c++1y")
-        .compile("solver_cpp");
-
     Ok(())
 }
